backend/flaskr/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
redis
flask-cors
docker
mysql-connector-python
pymysql
numpy
bandit
flask-mail
itsdangerous
flask_migrate
pytest
pytest-flask
locust